% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_adm_2.R
\name{agreguj_1rokpo_adm_2}
\alias{agreguj_1rokpo_adm_2}
\title{Obliczanie wskaznikow na poziomie zagregowanym dla 2. edycji
monitoringu na danych administracyjnych}
\usage{
agreguj_1rokpo_adm_2(wsk2, wsk3, wsk4, grupy, rok, duplikaty = TRUE)
}
\arguments{
\item{wsk2}{ramka danych z tabeli pośredniej nr 2 (P2) z wynikami z 2.
rundy monitoringu na danych administracyjnych}

\item{wsk3}{ramka danych z tabeli pośredniej nr 3 (P3) z wynikami z 2.
rundy monitoringu na danych administracyjnych}

\item{wsk4}{ramka danych z tabeli pośredniej nr 4 (P4) z wynikami z 2.
rundy monitoringu na danych administracyjnych}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}

\item{rok}{rok, w którym grupa absolwentów uzyskała status absolwenta. Może
to być więcej niż 1 wartość.}

\item{duplikaty}{wartość logiczna określająca czy należy odfiltrować ze
zbioru duplikaty przy liczeniu niektórych wskaźników. Przyjmuje wartości:
\itemize{
\item{TRUE}{wartość domyślna; duplikaty nie zostaną odfiltrowane ze zbioru}
\item{FALSE}{duplikaty zostaną odfiltrowane ze zbioru}
}}
}
\value{
data frame
}
\description{
Funkcja obliczająca wskaźniki na poziomie zagregowanym na
potrzeby raportów "1rokpo"
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{dane_szkoly}},}
\item{\code{\link{l_abs}},}
\item{\code{\link{l_kobiet}},}
\item{\code{\link{l_abs_zrodla}},}
\item{\code{\link{S3_mies}},}
\item{\code{\link{zawody_S3}},}
\item{\code{\link{E2_nauka_kontyn}},}
\item{\code{\link{liczebnosc_branze_ucz}},}
\item{\code{\link{liczebnosc_branze_kont}},}
\item{\code{\link{Z4_ods_prac_mies}},}
\item{\code{\link{Z8_formy_prac_mies}},}
\item{\code{\link{Z9_kont_mlod}},}
\item{\code{\link{W3_sr_doch_uop}},}
\item{\code{\link{B2_ods_bezrob}},}
\item{\code{\link{N2_ods_biernosc}},}
\item{\code{\link{licz_zawody}},}
\item{\code{\link{liczebnosc_dyscypliny}},}
\item{\code{\link{liczebnosc_dziedziny}},}
\item{\code{\link{dyscypliny_zawody}}}
}
}
