% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{branze_zawody}
\alias{branze_zawody}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
branze_zawody(x, branza_kont_df, rok, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{branza_kont_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dyscyplinie (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie
grudzień}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę na studiach w podziale na dyscypliny i zawody - wynik działania funkcji
jest wsadem do tabeli krzyżowej dyscypliny przez zawody w raporcie. Funkcja
liczy wskaźnik tylko dla absolwentów techników i liceów ogólnokształcących.
Wskaźnik liczony jest tylko dla zawodów, w których uczyło się więcej niż 10
absolwentów (n>=10).
}
