% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{data_na_okres}
\alias{data_na_okres}
\title{Funkcje robocze dla obliczania wskaźników dla 2. edycji monitoringu -
dane administracyjne}
\usage{
data_na_okres(mies, rok)
}
\arguments{
\item{mies}{miesiąc wyrażony jako liczba z zakresu od 1 do 12, gdzie styczeń
to wartość 1, luty to wartość 2 itd.}

\item{rok}{rok kalendarzowy wyrażony jako liczba}
}
\value{
numeric
}
\description{
Funkcja przeliczająca miesiąc i rok na okres, czyli liczbę
miesięcy od roku 0
}
