% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{E2_nauka_kontyn}
\alias{E2_nauka_kontyn}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
E2_nauka_kontyn(x, rok, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik E2 - Sposoby kontynuowania edukacji.
}
