% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{l_kobiet}
\alias{l_kobiet}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
l_kobiet(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę kobiet wśród
absolwentów na danym poziomie agregacji, którzy zostali objęci monitoringiem.
}
