% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{licz_zawody}
\alias{licz_zawody}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
licz_zawody(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności w zawodach na potrzeby
raportu wojewódzko-branżowego.
}
