% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{liczebnosc_branze_kont}
\alias{liczebnosc_branze_kont}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
liczebnosc_branze_kont(x, branza_kont_df, rok, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{branza_kont_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej branży (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie
grudzień}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę w szkołach branżowych 2. stopnia w podziale na branże. Dodatkowo,
funkcja liczy liczebności absolwentów w branżach, a informacja ta służy jako
podstawa do definiowania warunków w szablonie raportu dla szkół branżowych 1.
stopnia. Funkcja ma sens tylko dla absolwentów szkół branżowych 1. stopnia,
ponieważ tylko oni mogą kontynuować naukę w szkołach branżowych 2. stopnia.
}
