% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{liczebnosc_branze_ucz}
\alias{liczebnosc_branze_ucz}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
liczebnosc_branze_ucz(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności wyuczonych zawodów wśród
absolwentów w podziale na branże. Dodatkowo, funkcja liczy liczebności
absolwentów w branżach, a informacja ta służy jako podstawa do definiowania
warunków w szablonie raportu dla szkół branżowych 1. stopnia.
}
