% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_robocze.R
\name{okres_na_date}
\alias{okres_na_date}
\title{Funkcje robocze dla obliczania wskaźników dla 2. edycji monitoringu -
dane administracyjne}
\usage{
okres_na_date(x)
}
\arguments{
\item{x}{wartość liczbowa, która ma być przeliczona na miesiąc i rok
(zazwyczaj jest to zmienna \code{OKRES} w tabeli pośredniej)}
}
\value{
list
}
\description{
Funkcja przeliczająca liczbę miesięcy, które minęły od roku 0 na
miesiąc i rok.
}
