% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{podregiony}
\alias{podregiony}
\title{Mapowanie kodów TERYT powiatów na podregiony statystyczne (wg NUTS
2016)}
\format{
Ramka danych z 2 kolumnami i 380 wierszami.
}
\usage{
podregiony
}
\description{
Ramka danych z sześciocyfrowymi (lub pięciocyfrowymi) kodami
TERYT powiatów oraz nazwami podregionów statystycznych, do których
poszczególne powiaty są przypisane
}
\keyword{datasets}
