% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiczna_grupa_odniesienia.R
\name{utworz_grupowanie_odn_teryt}
\alias{utworz_grupowanie_odn_teryt}
\title{Obliczanie wskaznikow na poziomie zagregowanym - definiowanie grupowania}
\usage{
utworz_grupowanie_odn_teryt(
  x,
  zmGrupujace,
  ...,
  zmPominGrupaOdniesienia = any_of("id_szk"),
  idAbs = all_of(c("id_abs", "rok_abs")),
  zmTerytPow = all_of("teryt_pow_szk"),
  minN = 10,
  minM = 3,
  wykluczGrupeZGrupyOdniesienia = TRUE
)
}
\arguments{
\item{x}{ramka danych z danymi indywidualnymi (zwykle \emph{tabela pośrednia}
P4)}

\item{zmGrupujace}{nazwa lub wektor nazw zmiennych (\emph{\<tidy-select\>}),
kombinacja wartości których definiuje podział na grupy}

\item{...}{opcjonalnie specyfikacja dodatkowych kolumn z ramki danych
podanej argumentem \code{x}, które mają zostać dołączone do zwracanej ramki
danych (\emph{\<tidy-select\>})}

\item{zmPominGrupaOdniesienia}{nazwa lub wektor nazw zmiennych
(\emph{\<tidy-select\>}), które należy \strong{pominąć} względem zbioru
podanego w argumencie \code{zmGrupujace}, przy definiowaniu przypisania do
\emph{grupy odniesienia} (pomijając kryterium geograficzne, które zostanie
obsłużone oddzielnie)}

\item{idAbs}{nazwa lub wektor nazw zmiennych (\emph{\<tidy-select\>}),
kombinacja wartości których stanowi identyfikator absolwenta}

\item{zmTerytPow}{nazwa zmiennej zawierającej kod powiatu (w którym znajduje
się grupa, do której należy dany absolwent) (\emph{\<tidy-select\>})}

\item{minN}{minimalna liczba absolwentów, jaką musi obejmować grupa
odniesienia (jeśli \code{wykluczGrupeZGrupyOdniesienia=FALSE}, po odjęciu
liczby absolwentów w danej grupie)}

\item{minM}{minimalna liczba innych grup, jaką musi obejmować grupa
odniesienia (jeśli \code{wykluczGrupeZGrupyOdniesienia=FALSE}, powiększona
o 1 względem podanej)}

\item{wykluczGrupeZGrupyOdniesienia}{wartość logiczna - czy absolwenci
z analizowanej grupy powinni zostać wykluczeni z grupy odniesienia}
}
\value{
ramka danych, która może zostać użyta jako argument \code{grupy}
w wywołaniu funkcji \code{\link{agreguj_wskazniki}}.
}
\description{
Funkcja pozwala przygotować ramkę danych opisującą podział na
grupy, dla których mają następnie zostać obliczone wskaźniki zagregowane
przy pomocy funkcji \code{\link{agreguj_wskazniki}}, oraz odpowiadające im
grupy odniesienia w sytuacji, kiedy te drugie mają zostać utworzone w oparciu
o kryterium geograficzne. Jako geograficzna jednostka odniesienia wybierany
jest powiat, podregion statystyczny, województwo lub cała Polska w zależności
od tego, która z nich spełnia zadane argumentami \code{minN} i \code{minM}
kryteria minimalnej liczby odpowiednio absolwentów i grup. Jeśli kryteria te
nie są spełnione nawet dla województwa, jako kod specyfikujący grupę
odniesienia zwracany jest brak danych.
}
\examples{
\dontrun{
# wywołanie definiujące grupy na poziomie szkoło-zawodo-lat
szkoloZawody <- utworz_grupowanie_odn_teryt(
  p4,
  c(id_szk, kod_zaw, rok_abs),
  nazwa_zaw, branza)
# wywołanie definiujące grupy na poziomie szkoło-lat
szkoly <- utworz_grupowanie_odn_teryt(
  p4,
  c(id_szk, rok_abs))
}
}
