% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamiczna_grupa_odniesienia.R
\name{utworz_warunki}
\alias{utworz_warunki}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
utworz_warunki(x)
}
\arguments{
\item{x}{ramka danych zawierająca zmienne, kombinacje wartości których
definiują przypisanie do grupy (lub grupy odniesienia)}
}
\value{
wektor tekstowy
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{utworz_grupowanie_odn_teryt}} do sklejania ciągów znaków
zawierających wyrażenia opisujące kryteria przynależności do grup lub grup
odniesienia
}
\seealso{
\code{\link{utworz_grupowanie_odn_teryt}}
}
