% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{W3_sr_doch_uop}
\alias{W3_sr_doch_uop}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
W3_sr_doch_uop(x, rok, od = 9, do = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{od}{początek okresu wyliczania wskaźnika wyrażony miesiącem}

\item{do}{koniec okresu wyliczania wskaźnika wyrażony miesiącem}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
lista
}
\description{
Funkcja licząca średni względny miesięczny przychód z pracy
\emph{etatowej} w danym okresie w odniesieniu do zarobków w zamieszkiwanym
powiecie. W raporcie w 2021 roku będzie okres od \emph{września 2020} do
\emph{grudnia 2020} roku. Średnia liczona jest oddzielnie dla uczących i nie
uczących się absolwentów.
}
