% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{Z8_formy_prac_mies}
\alias{Z8_formy_prac_mies}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
Z8_formy_prac_mies(x, rok, mies = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów wykonujących dane formy
pracy w danym miesiącu (w raportach jest to grudzień). Wskaźnik może być
liczony albo dla absolwentów pracujących i kontunuujących naukę (\code{nauka
= TRUE}) lub dla absolwentów pracujących i nie kontunuujących nauki
(\code{nauka = FALSE}). Formy pracy:
\itemize{
\item{\code{ucz_uop} - Zatrudnienie na podstawie UOP oraz brak innej formy},
\item{\code{ucz_samoz} - Prowadzący działalność gosp. (samozatrudnienie)
oraz brak innej formy},
\item{\code{ucz_inna} - Tylko forma inna niż samozatrudnienie i niż umowa o
pracę (np. umowa-zlecenie)},
\item{\code{ucz_wiecej} - Więcej niż jedna forma, czyli: Zatrudnienie na
podstawie UOP oraz prowadzący działalność gosp., Zatrudnienie na podstawie
UOP oraz zatrudnienie w innej formie, Prowadzący działalność gosp.
(samozatrudnienie) oraz zatrudnienie w innej formie, Zatrudnienie na
podstawie UOP oraz działalność gosp. oraz zatrudnienie w innej formie}
}
}
