% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{Z9_kont_mlod}
\alias{Z9_kont_mlod}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
Z9_kont_mlod(x, rok, mies = 9)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów, którzy w danym miesiącu (w
raporcie jest to wrzesień) kontynuowali zatrudnienie u pracodawcy, który
wcześniej (kiedy byli jeszcze uczniami) zatrduniał ich jako pracowników
młodocianych. Wskaźnik liczony jest tylko dla absolwentów branżowych szkół 1.
stopnia. Wyróżnianych jest 5 form kontynuowania pracy u pracodawcy
zatrudniającego poprzednio jako młodocianego:
\itemize{
\item{\code{nieucz_niekontuop} - Nieuczący się, nie kontynuujący pracy u
danego pracodawcy}
\item{\code{nieucz_kont_uop} - Nieuczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{nieucz_kont_inne} - Nieuczący się, kontynuujący pracę u danego
pracodawcy w formie innej niż umowa  o pracę}
\item{\code{ucz_niekontuop} - Uczący się, nie kontynuujący pracy u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{ucz_kontuop} - Uczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
}
}
