% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_adm_2.R
\name{zawody_S3}
\alias{zawody_S3}
\title{Obliczanie wskaźników dla 2. edycji monitoringu - dane administracyjne}
\usage{
zawody_S3(x, rok_od, mies_od, rok_do, mies_do, dup = NULL)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{dup}{ciąg id absolwentów, których kombinacja zmiennych \code{id_abs}
oraz \code{rok_abs} występuje więcej niż raz w tabeli P4 i które mają zostać
odfiltrowane \strong{lub} wartość \code{NULL} jeśli obserwacje nie muszą być
odfiltrowywane ze zbioru (wartość domyślna)}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym wśród grup absolwentów wydzielonych ze względu na zawód,
w którym się kształcili w szkole. Zwracana lista służy jako wsad do tabeli w
raporcie automatycznym. Domyślnie funkcja ma zwracać wsad do tabeli dla
grudnia 2020 lub 2021, ale można to zmienić poprzez zmianę argumentów
funkcji. Jeśli lista będąca wsadem tabeli generowałaby pustą tabelę, to
zwracana jest pusta lista.
}
\details{
Funkcja wykorzystuje inną funkcję z pakietu, czyli
\code{\link{S3_mies}}
}
